Set DotNetUseShippingVersions
This fixes a s390x native build issue where
Microsoft.Interop.LibraryImportGenerator would fail to load with an
obscure error.

The build uses two versions of this assembly: one shipped with the host
SDK, and one produced by the build itself. When the host SDK one is
properly versioned, and the built one has a non-shipping version
42.42.42.42 [1], the built one takes priority [2]. When they both have
version 42.42.42.42, the version check gets confused.

There seems to be a deficiency in DefaultAnalyzerAssemblyLoader design.
It is owned by CompilerServerHost and can be shared between multiple
compilations. DefaultAnalyzerAssemblyLoader maintains a mapping between
simple assembly names and the respective .dll paths, which is populated
based on /analyzer compiler flag. The values from the old compilations
are not erased. This leads to the following sequence:

* Compilation A requests host Microsoft.Interop.LibraryImportGenerator
  and its dependency - Microsoft.Interop.SourceGeneration.dll. As
  required by design, it passes paths to both with the /analyzer flag.
  These paths are stored in _knownAssemblyPathsBySimpleName.

* Compilation B requests the built versions of these assemblies. There
  are two copies of SourceGeneration: one under
  artifacts/bin/Microsoft.Interop.SourceGeneration, and another one
  under artifacts/bin/LibraryImportGenerator. Compilation requests the
  first one [3]. The respective paths are appended to
  _knownAssemblyPathsBySimpleName.

* Three assembly loads occur. Two are triggered by the /analyzer flag
  and complete correctly, because the `paths.Contains(assemblyPath)`
  check trivially succeeds.

* The third one loads SourceGeneration in order to satisfy the
  reference from LibraryImportGenerator. Therefore it uses the latter's
  ALC, which is DirectoryLoadContext(Directory=
  artifacts/bin/LibraryImportGenerator). Therefore assemblyPath
  corresponds to the second copy of SourceGeneration, which is not in
  _knownAssemblyPathsBySimpleName. Therefore we proceed to the version
  check loop, which SourceGeneration loaded by compilation A trivially
  wins.

The quick hack here is to make it lose by assigning a proper version to
it.

I believe a proper solution is to unshare
DefaultAnalyzerAssemblyLoader. Clearing _knownAssemblyPathsBySimpleName
between compilations is not feasible, since
DefaultAnalyzerAssemblyLoader may be used by multiple threads.

[1] https://github.com/dotnet/arcade/blob/23681df57e8b3d551e5ab2cab3e8e1e5e9d6f35f/src/Microsoft.DotNet.Arcade.Sdk/tools/Version.targets#L36
[2] https://github.com/dotnet/roslyn/blob/f6f1c7171ce27ad23fc3c87d1c996dedf2156f44/src/Compilers/Core/Portable/DiagnosticAnalyzer/DefaultAnalyzerAssemblyLoader.Core.cs#L161
[3] https://github.com/dotnet/runtime/blob/3b706dd67bd0d5a29a5bf1801e970b94e0d0f33e/eng/generators.targets#L40